requirements.txt
discord.py==1.7.3
webcolors==1.12
